add language to index
